Refactor to SensorService and fix the optional variable.
